Avance en código tp1 y armado de clase para los test

+++ src/ar/edu/unq/po2/tp2/EmpleadoPlantaPerm.java
@@ -51,13 +51,13 @@ public class EmpleadoPlantaPerm extends Empleado{
 		return this.calcularBruto() - this.calcularRetenciones();
 	}
 	
-	public EmpleadoPlantaPerm(String nombre, String direccion, String estadoCivil, float fechaDeNacimiento,
-			                  float sueldoBasico, float cantHijos, float antiguedad) {
+
+	public EmpleadoPlantaPerm(String nombre, String direccion, String estadoCivil, int fechaDeNacimiento,
+			float sueldoBasico, double cantHijos, double antiguedad) {
 		super(nombre, direccion, estadoCivil, fechaDeNacimiento, sueldoBasico);
 		this.cantHijos = cantHijos;
 		this.antiguedad = antiguedad;
 	}
-
 
 	public double getCantHijos() {
 		return cantHijos;
